+++ pyaerocom/__init__.py
@@ -53,7 +53,7 @@ from .time_resampler import TimeResampler
 from .io.helpers import search_data_dir_aerocom
 from .variable import get_variable
 from .utils import create_varinfo_table
-from .testdata_access import initialise as initialise_testdata
+from .access_testdata import initialise as initialise_testdata
 
 from . import aeroval
