+++ pyaerocom/access_testdata.py
